backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.4.2
google-cloud-secret-manager==2.16.4
python-multipart==0.0.6